+++ src/vq/vq_learn.c
@@ -10,6 +10,7 @@ int vq_learn(int prediction_order,
              const char *predictor_filenames[],
              int num_predictors,
              int use_par,
+             void* callback_target,
              vq_learn_callback_t callback
         ) {
 
@@ -25,6 +26,7 @@ int vq_learn(int prediction_order,
             predictor_filenames,
             num_predictors,
             use_par,
+            callback_target,
             callback
             );
 }
@@ -35,6 +37,7 @@ int vq_learn_using_base_codebook(
         const char *predictor_filenames[],
         int num_predictors,
         int use_par,
+        void* callback_target,
         vq_learn_callback_t callback
                                 ) {
 
@@ -64,6 +67,7 @@ int vq_learn_using_base_codebook(
             predictor_filenames,
             num_predictors,
             use_par,
+            callback_target,
             callback
     );